% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GoogleAPI.R
\name{GoogleAPI}
\alias{GoogleAPI}
\title{R interface for Google Civic API}
\usage{
GoogleAPI(method, key = getOption("GoogleAPIKey"), debug = FALSE)
}
\arguments{
\item{method}{method to call}

\item{key}{API key}

\item{debug}{Enable debugging mode}
}
\description{
See \url{https://developers.google.com/civic-information/}
}
\seealso{
\code{\link{PopongAPI}}, \code{\link{SunlightAPI}}
}
\keyword{API}

